requirements.txt
asgiref==3.11.1
asttokens==3.0.1
crispy-tailwind==1.0.3
decorator==5.2.1
Django==6.0.3
django-crispy-forms==2.6
djangorestframework==3.17.1
executing==2.2.1
ipython==9.12.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.2.1
parso==0.8.6
pexpect==4.9.0
pillow==12.2.0
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.20.0
sqlparse==0.5.5
stack-data==0.6.3
traitlets==5.14.3
wcwidth==0.6.0
